Add pyframebuffer as buildroot python package
Adds pyframebuffer (https://github.com/RossAdrian/pyframebuffer) as external python module to buildroot. As like other third-party python modules in buildroot, you can just select it, and it will be build and installed on the target system. All tests to build this package with buildroot under usual conditions have been succeeded, as well as it's usage in the embedded system in Qemu.